forwarder: implement replayable transactions.
Sending data to the Datadog intake is now done through a "Transaction" object
container the HTTP request payload. If sending this data fails, the transaction
is stored within the forwarder, and will be replayed when a next transaction
eventually succeed. After some tries, a transaction is automatically dropped.

A next iteration on this would be to not limit "how many transactions" we store
but to limit the amount of transactions based on "how much ram is used" by these
stored transactions. This would be a configuration field.