sampler: copy tags from parser to sampler.
In order to be able to delete the whole arena memory used by the parser, we have
to copy the tags it had allocated into the sampler allocator otherwise, the tags
can be dropped (i.e. the arena can be dropped) while we would still need them
for the serialization.